Rename main.py to app.py
